+++ plugins/kiali/src/store/KialiProvider.tsx
@@ -17,6 +17,7 @@ import { IstioStatusActions } from '../actions/IstioStatusActions';
 import { MeshTlsActions } from '../actions/MeshTlsActions';
 import { setServerConfig } from '../config/ServerConfig';
 import { KialiHelper } from '../pages/Kiali/KialiHelper';
+import { KialiNoResources } from '../pages/Kiali/KialiNoResources';
 import {
   HelpDropdownStateReducer,
   IstioCertsInfoStateReducer,
@@ -59,7 +60,9 @@ export const KialiProvider: React.FC<Props> = ({
   const promises = new PromisesRegistry();
   const [kialiCheck, setKialiCheck] =
     React.useState<KialiChecker>(initialChecker);
-
+  const [notHaveResources, setNotHaveResources] = React.useState<
+    boolean | undefined
+  >(undefined);
   const [loginState, loginDispatch] = React.useReducer(
     LoginReducer,
     initialStore.authentication,
@@ -106,6 +109,11 @@ export const KialiProvider: React.FC<Props> = ({
           namespaceDispatch(
             NamespaceActions.receiveList([...data], new Date()),
           );
+          if (data.length > 0) {
+            setNotHaveResources(false);
+          } else {
+            setNotHaveResources(true);
+          }
           namespaceDispatch(NamespaceActions.setActiveNamespaces([...data]));
         })
         .catch(() => namespaceDispatch(NamespaceActions.requestFailed()));
@@ -221,7 +229,13 @@ export const KialiProvider: React.FC<Props> = ({
   );
   useDebounce(refresh, 10);
 
-  return !loading ? (
+  if (loading) {
+    return <CircularProgress />;
+  } else if (!!notHaveResources) {
+    return <KialiNoResources entity={entity!} />;
+  }
+
+  return (
     <KialiContext.Provider
       value={{
         authentication: loginState,
@@ -243,7 +257,5 @@ export const KialiProvider: React.FC<Props> = ({
     >
       {kialiCheck.verify ? children : <KialiHelper check={kialiCheck} />}
     </KialiContext.Provider>
-  ) : (
-    <CircularProgress />
   );
 };