+++ plugins/orchestrator-common/src/index.ts
@@ -1,4 +1,6 @@
 export * from './types';
+export * from './openapi/types';
+export * from './auto-generated/api/definition';
 export * from './constants';
 export * from './models';
 export * from './workflow';